decouple VisibleBreathing and Hypothermia
