+++ src/main/java/org/terasology/climateConditions/HypothermiaSystem.java
@@ -16,8 +16,6 @@
 package org.terasology.climateConditions;
 
 import org.terasology.entitySystem.entity.EntityRef;
-import org.terasology.entitySystem.entity.lifecycleEvents.BeforeRemoveComponent;
-import org.terasology.entitySystem.entity.lifecycleEvents.OnAddedComponent;
 import org.terasology.entitySystem.event.ReceiveEvent;
 import org.terasology.entitySystem.systems.BaseComponentSystem;
 import org.terasology.entitySystem.systems.RegisterMode;
@@ -26,20 +24,12 @@ import org.terasology.logic.characters.AffectJumpForceEvent;
 import org.terasology.logic.characters.CharacterMovementComponent;
 import org.terasology.logic.characters.GetMaxSpeedEvent;
 import org.terasology.logic.characters.MovementMode;
-import org.terasology.logic.delay.DelayManager;
 import org.terasology.logic.players.PlayerCharacterComponent;
 import org.terasology.physics.events.MovedEvent;
-import org.terasology.registry.In;
 
 @RegisterSystem(value = RegisterMode.AUTHORITY)
 public class HypothermiaSystem extends BaseComponentSystem {
-
-    @In
-    private DelayManager delayManager;
-
     private float thresholdHeight = 60f;
-    private int initialDelay = 5000;
-    private int breathInterval = 7000;
     private float walkSpeedMultiplier = 0.6f;
     private float jumpSpeedMultiplier = 0.7f;
 
@@ -56,16 +46,6 @@ public class HypothermiaSystem extends BaseComponentSystem {
                 player.removeComponent(HypothermiaComponent.class);
             }
         }
-    }
-
-    @ReceiveEvent(components = {HypothermiaComponent.class})
-    public void onHypothermia(OnAddedComponent event, EntityRef player, HypothermiaComponent hypothermia) {
-        delayManager.addPeriodicAction(player, VisibleBreathingSystem.VISIBLE_BREATH_ACTION_ID, initialDelay, breathInterval);
-    }
-
-    @ReceiveEvent(components = HypothermiaComponent.class)
-    public void beforeRemoveHypothermia(BeforeRemoveComponent event, EntityRef player, HypothermiaComponent hypothermia) {
-        delayManager.cancelPeriodicAction(player, VisibleBreathingSystem.VISIBLE_BREATH_ACTION_ID);
     }
 
     @ReceiveEvent(components = {HypothermiaComponent.class})
